+++ test/unit/fuzz-tools.test.js
@@ -466,6 +466,7 @@ import * as rockwoodAcV487 from '../../lib/rockwood-ac-v487.js';
 import * as biglianiAcromionV488 from '../../lib/bigliani-acromion-v488.js';
 import * as fernandezRadiusV489 from '../../lib/fernandez-radius-v489.js';
 import * as ruediAllgowerPilonV490 from '../../lib/ruedi-allgower-pilon-v490.js';
+import * as severinDdhV491 from '../../lib/severin-ddh-v491.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -877,6 +878,7 @@ const MODULES = {
   'bigliani-acromion-v488.js': biglianiAcromionV488,
   'fernandez-radius-v489.js': fernandezRadiusV489,
   'ruedi-allgower-pilon-v490.js': ruediAllgowerPilonV490,
+  'severin-ddh-v491.js': severinDdhV491,
   'tb-testing.js': tbTesting,
 };